Migrate logic from `node.rs` into `mod.rs` following the removal of RenderGraph.
