+++ crates/bevy_anti_alias/src/taa/mod.rs
@@ -17,7 +17,7 @@ use bevy_ecs::{
 };
 use bevy_image::{BevyDefault as _, ToExtents};
 use bevy_math::vec2;
-use bevy_post_process::{bloom::bloom, motion_blur::node::motion_blur};
+use bevy_post_process::{bloom::bloom, motion_blur::motion_blur};
 use bevy_reflect::{std_traits::ReflectDefault, Reflect};
 use bevy_render::{
     camera::{ExtractedCamera, MipBias, TemporalJitter},
